refactor(alarm): sleep 모드 가능 여부
- 이전 코드의 비즈니스적 오류 다수 발생 및 좋지 못한 코드퀄리티로 일괄 변경함
- 최종적으로 66줄 로직은 단 4줄로 처리함

+++ BE/src/main/java/com/sf/honeymorning/alarm/dto/request/AlarmSetRequest.java
@@ -17,7 +17,7 @@ public record AlarmSetRequest(
 	@Min(1)
 	@Max(127)
 	@NotNull
-	byte weekdays,
+	Integer weekdays,
 
 	@Positive
 	@Min(1)
